+++ NEWS.md
@@ -9,6 +9,7 @@ Minimum LaBB-CAT version *20211117.2027*
 - *updateFragment*: Uploads an uttereance file (e.g. TextGrid) to update labels and/or alignments.
 - *generateLayer*: Generates annotations on a given layer for all transcripts in the corpus.
 - *generateLayerUtterances*: Generates a layer for a given utterances.
+- *annotatorExt*: Retrieves an annotator extension resource.
 
 # nzilbb.labbcat 1.0-1